Question 9
I updated question 9 to accept an Account Name and then queried to see
if the Account existed and if any records found were added them to a List.
If the Account existed I updated the Account Description. If the Account
did not exist I created the Account and then added it to the List.
Then I upserted the list of Accounts.