Question 10
I created a List to store the Last Name of each Contact which were going
to be used as the Account Name. I then looped through each Contact in the
contactsToUpsert List and added each Contacts Last Name to my List.
Next I ran a SOQL Query to locate each Account if it existed and stored
those Accounts. I then checked each name in my List against the name of
each account and if not found I created a new account and added it to the
 list of queried accounts. I then upserted the Accounts List to ensure
 each account would have an Id that could be related to each contact.
 Then I looped through each contact and compared it to each account in
 the and when found assigned that Account Id to the Contact. Lastly I
 upserted all the Contacts with their related AccountId.